pyproject.toml
[tool.poetry]
name = "drawme"
version = "0.1.0"
description = ""
authors = ["Gabri Botha <Angelodeath@outlook.com>"]

[tool.poetry.dependencies]
python = "^3.7"
pillow = "^7.2.0"
pygame = "^1.9.6"
evdev = "^1.3.0"

[tool.poetry.dev-dependencies]
pytest = "^5.2"
black = "^19.10b0"

[build-system]
requires = ["poetry>=0.12"]
build-backend = "poetry.masonry.api"
